refactor(furnacemind): move feedback service modules
