--- src/agents/feedback/__init__.py
@@ -1,1 +1,0 @@
-"""Feedback detection, persistence, and lesson retrieval for agents."""
