+++ src/utils/furnacemind/__init__.py
@@ -1,0 +1,1 @@
+"""FurnaceMind utility services."""
